src: remove multiple uses of atoi
atoi() and related functions lack a mechanism for reporting errors for
invalid values. Replace them with calls to the appropriate
ByteExtractString* functions.

Partially closes redmine ticket #3053.

+++ src/source-netmap.h
@@ -48,7 +48,7 @@ typedef struct NetmapIfaceSettings_
     bool ips;       /**< set to true if checksum_mode != NETMAP_COPY_MODE_NONE */
     bool threads_auto;
 
-    int threads;
+    uint16_t threads;
     int copy_mode;
     ChecksumValidationMode checksum_mode;
     const char *bpf_filter;
